fix sizing handling of chart
