+++ packages/react-widgets/src/widgets/TimeSeriesWidget.js
@@ -125,6 +125,7 @@ function TimeSeriesWidget({
   onTimeWindowUpdate,
   onStateChange,
   palette,
+  showLegend,
   // Both
   stepSize,
   stepMultiplier,
@@ -353,6 +354,7 @@ function TimeSeriesWidget({
               onSelectedCategoriesChange={handleSelectedCategoriesChange}
               isLoading={isLoading}
               palette={palette}
+              showLegend={showLegend}
             />
           )}
         </WidgetWithAlert>
